backend/requirements.txt
Pillow==10.1.0
python-jose[cryptography]==3.5.0